pyproject.toml
[project]
name = "async-rl-lab"
version = "0.1.0"
description = "Modular async GRPO trainer for controlled RL experiments"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "safetensors>=0.4.0",
    "datasets>=2.18.0",
    "accelerate>=0.28.0",
    "vllm>=0.6.0",
    "wandb>=0.15.0",
    "matplotlib>=3.8.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
redis = ["redis>=5.0.0"]
zmq = ["pyzmq>=25.0.0"]
all = ["redis>=5.0.0", "pyzmq>=25.0.0"]

[tool.hatch.build.targets.wheel]
packages = ["core", "buffers", "weight_sync", "staleness", "interrupts", "scorers", "utils"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

